tablas oracle terminado (revisar dates)

+++ db/migrations/20250607223532_create_tarjetas_corredor.sql
@@ -4,7 +4,7 @@ CREATE TABLE tarjetas_corredor (
   id INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
   id_tarjeta_fisica INTEGER NOT NULL,
   saldo FLOAT NOT NULL,
-  id_perfil FLOAT NOT NULL,
+  id_perfil INTEGER NOT NULL,
   FOREIGN KEY (id_perfil) REFERENCES perfiles (id)
 );
